Ref(app.*): Moving toolbar and files to own export component
