+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { NotFoundComponent } from './pages/not-found.component';
-import { HomeComponent } from './pages/home.component';
+import { NotFoundComponent } from './core';
+import { HomeComponent } from './core';
 
 const routes: Routes = [
   {
