+++ src/app/app.component.scss
@@ -2,6 +2,6 @@
 :host ::ng-deep .p-menubar {
     border: 0;
     height: 60px;
-    
+    color: aqua;
+
 }
-
